refactor: 멤버 정보 DTO 및 기반 서비스 컨트롤러 수정
