+++ src/test/java/com/example/masterplanbbe/member/api/MemberApiTest.java
@@ -1,8 +1,6 @@
 package com.example.masterplanbbe.member.api;
 
-import com.example.masterplanbbe.member.service.request.MemberCreateRequest;
-import com.example.masterplanbbe.member.service.request.MemberUpdateRequest;
-import com.example.masterplanbbe.member.service.response.MemberResponse;
+import com.example.masterplanbbe.member.dto.MemberResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.client.RestClient;
 import lombok.AllArgsConstructor;
